Removed redundancy in initialization
I removed the initialize.py script that was redundant with the
initial_data.json and also cleaned up the initial_data.json so it
corresponded to what was in the initialize.py script so that things
would work.

I have done basic testing (wiped, reinitialized, poked around) but not
any automated testing because I can't get automated testing to work
(!!!).